Tests - move core mocks to support directory
